Warn when a mock is used as a context manager
Instead of raising an exception, warn when a pytest-mock mock is used as
a context manager. This lets pytest-mock mock objects that are used as
context managers, like threading.Lock, while still letting users know
that mocks returned from pytest-mock do not need their __enter__ method
called to take effect.

Create a PytestMockWarning class so that the warning is easy to
pinpoint and handle.