autospec for spy: remove if else which was in place for python < 3.7 (#399)
The if/else for the spy feature was there to get around a bug in python that was fixed in 3.7. 

Now that we only support >= 3.8 this check is no longer needed.

This should not really affect users, but if we learn it does, we can add a CHANGELOG later (as discussed in #399).